requirements.txt
numpy
Pillow
torch
scikit-learn
pyyaml
